Refactor: Move the callbacks to concern
